refactor(lightbeam): author every board, and retire route-then-obstruct
Completes the migration. `LIGHTBEAM_CONFIG` now holds the authored tier table and
every board a player meets is built by authoring its maze: lay a golden path, then
author a dying corridor for every stop a mirror is not set to, recursing wherever a
corridor meets a piece whose state the beam has not fixed.

Gone with it: `buildRoute`, `blockWrongSettings`, `placeShadows`, `mirrorStopSet`,
`thinWalls`, the goal pool (`GOAL_DIALS`, `drawGoals`, `LIGHTBEAM_GOALS` and the
fallback ladder), and the `goals` field on a puzzle. Modes replace the goals, and a
board records the modes it was built to for the same reason it recorded its goals:
a fallback that fires quietly makes the whole pool decorative while every other
measurement still looks fine.

What each tier now gets, measured over 40 seeds, against what it got before:

  tier     pieces      configs      rejects/board   worst gen
  starter  3.0 -> 3.0    8 ->    8   1.3 ->   0.0   14ms -> 10ms
  junior   4.0 -> 4.0   16 ->   16   3.0 ->   0.1   16ms ->  7ms
  expert   5.9 -> 5.5  109 ->   80  70.5 ->   2.3   26ms -> 20ms
  master   7.0 -> 7.0  230 ->  228 355.7 ->   1.1   45ms -> 19ms
  wizard   8.3 -> 9.8  934 -> 1741 226.0 ->   1.7  550ms -> 616ms

Two gaps adopted knowingly: expert is slightly thinner than the tier it replaces
while demanding deeper rungs, and wizard is the one tier that is slower, because
onlySurvivor enumerates the whole product on a bigger board. That is now the last
measured bottleneck in the family and it is written into §11 Deferred with the fix.

Structure: `lightbeamGeometry.ts` holds what a board is made of — the mirror law,
rule 2's stop sets, §5.2's crossings, the spacing rule, the opening rules — and
`generateLightbeam.ts` holds how one is built. Nothing in the first knows about the
second. Neither file carries an "authored" qualifier, because there is no longer
another kind.

Five stale assertions in the spec were rewritten rather than deleted, and each was
stale for a reason worth keeping:

- Uniqueness is one winning ROUTE, not one winning configuration. A decoy's setting
  is free by construction, so any tier with a piece off the beam's line has many
  winning configurations.
- "Every wrong setting dies" holds only for the pieces the winning beam crosses.
  Asserted over all mirrors it says a decoy is not a decoy.
- Reuse is the design now, so only the join half of the pair invariant is asserted.
- A wall can be load-bearing only two wrong settings deep, so that check runs over
  every configuration rather than the single-piece deviations.
- Wall-heavy's corner pairs stop nothing on purpose; they are there to be read.

Docs: §5 describes authoring, §7 describes modes, §6's table is the measured tier
table with per-board configurations, and §6.4's catalogue of goal-pool mismatches is
replaced by what generation now refuses to build. §11 stays as the record of how it
was reached. `docs/lightbeam-authored-generation-implementation-plan.md` is deleted
as its own last instruction says, with its open items moved into §11 Deferred and
the three ways to measure this family wrongly into §11.0.

Verified in the browser: all five tiers build, open dark, redraw on a tap, and no
page errors. Ramp visible in the lab at 3 / 4 / 6 / 9 / 12 pieces. The lab's variant
picker now offers the three modes rather than two generators.

1848 tests green, 0 lint errors.
