+++ src/mods/puzzle/game/lightbeam/meta.ts
@@ -7,18 +7,9 @@ export const LIGHTBEAM_META: FamilyMeta = {
   // Configured for every tier (docs/game-design/puzzles/lightbeam.md §6), so the allocator may draw it
   // anywhere and a node may author it anywhere.
   minTier: "starter",
-  // Two generators coexist while the authored construction is measured against route-then-obstruct
-  // (design doc §11.16). The lab is the only place the second one is reachable; "shipped" is the default.
-  variants: [
-    "shipped",
-    "authored",
-    "authored tiers",
-    "authored wall-heavy",
-    "authored slider-heavy",
-    "authored switch-heavy",
-    "authored trap",
-    "authored all modes",
-  ],
+  // Modes a developer can force on top of a tier's own dials, to look at one shape at a time
+  // (docs/instructions/puzzle-screens.md §6). "tier default" is what real play uses.
+  variants: ["tier default", "wall-heavy", "slider-heavy", "switch-heavy"],
   icon: "🔆",
   color: "amber",
   rewardPriority: 60, // fills only once treasure's guaranteed slots are spoken for
